threat modeling agent/requirements.txt
langchain
langchain-core
python-dotenv
pydantic
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-google-vertexai
